% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccd.R
\name{fact.pts}
\alias{fact.pts}
\title{Generate Factorial Points}
\usage{
fact.pts(k)
}
\arguments{
\item{k}{an integer greater than
or equal to 2.}
}
\value{
For \code{K} number of factors, generate
factorial points with minimum level changes.
}
\description{
For a given number of factors, say k, it generates 2^k factorial
points such that the number of changes in the run sequence
is minimum.
}
\examples{
# For generating a full factorial for k=4 factors
# in 2 levels with minimum level changes
fact.pts(4)
}

