% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccd_halfrep.R
\name{fact.pts_2}
\alias{fact.pts_2}
\title{Generate Half Replicate of Full Factorial Points.}
\usage{
fact.pts_2(k)
}
\arguments{
\item{k}{an integer greter than or equal to 3.}
}
\value{
For \code{K} number of factors, generate
factorial points in half replication with minimum level changes.
}
\description{
For a given number of factors, say k, it generates 2^(k-1) factorial
points such that the number of changes in levels in the half
replicate is minimum.
}
\examples{
# For generating a minimally changed half
# replicate of full factorial with 4 factors
# in 2 levels fact.pts_2(4)
}

