% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbd.R
\name{min_bbd}
\alias{min_bbd}
\title{Minimally Changed BBD}
\usage{
min_bbd(k)
}
\arguments{
\item{k}{An integer greter than or equal to 3.}
}
\value{
returns a minimally changed BBD for the
 number of input factors as \code{v}.
}
\description{
Generate Box Behnken design (BBD) with minimum level changes
in the run sequence.
}
\examples{
# To generate minimmaly changed BBD for k=4 factors
min_bbd(4)
}

