% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccd.R
\name{min_ccd}
\alias{min_ccd}
\title{Minimally Changed CCD}
\usage{
min_ccd(k)
}
\arguments{
\item{k}{an integer greater than
or equal to 2.}
}
\value{
returns a minimally changed CCD for \code{K}
number of factors with full factorial points.
}
\description{
Generate minimally changed run sequence for central composite
designs (CCD).
}
\examples{
# Generate minimally changed CCD with full factorial points
# for k=4 factors
min_ccd(4)
}

