% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccd_halfrep.R
\name{min_ccd_2}
\alias{min_ccd_2}
\title{Generate Minimally Changed Central Composite
 designs (CCD) with Fractional Factorial Points (Half Replicate)}
\usage{
min_ccd_2(k)
}
\arguments{
\item{k}{An integer greter than or equal to 3.}
}
\value{
returns a minimally changed CCD for \code{K}
number of factors with half replicated factorial points.
}
\description{
Generate Minimally Changed Central Composite
 designs (CCD) with Fractional Factorial Points (Half Replicate)
}
\examples{
# Generate minimally changed CCD with fractional
# factorial points (half replicate) for k=4 factors
min_ccd_2(4)
}

