% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_changes.R
\name{num.ch}
\alias{num.ch}
\title{Number of Changes in Factor Levels of the Design}
\usage{
num.ch(M)
}
\arguments{
\item{M}{a design matrix.}
}
\value{
returns the number of changes in levels of each factor
and total number of such changes in the design
}
\description{
For a given design in matrix form, it counts the number of changes
in levels of each factor as well as overall number of changes in
the design.
}
\examples{
# For generating a minimally changed CCD with 4 factors
des<-min_ccd(4)
# To count the number of level changes in the generated design
num.ch(des$CCD_design)
}

